update collapsed lane
